split profile and subprofile editor in settings
