+++ src/components/Settings/Profile.tsx
@@ -21,7 +21,6 @@ import {
     type CoreProfile,
     type Schema,
     type BadgeRef,
-    type CoreTimeline,
     Schemas,
     type SubprofileTimelineSchema,
     type EmptyTimelineSchema,
@@ -33,23 +32,21 @@ import {
 import { useEffect, useState } from 'react'
 import { CCDrawer } from '../ui/CCDrawer'
 import { CCEditor } from '../ui/cceditor'
-import { SubProfileCard } from '../SubProfileCard'
 import ManageSearchIcon from '@mui/icons-material/ManageSearch'
 
-import DeleteForeverIcon from '@mui/icons-material/DeleteForever'
 import PublishIcon from '@mui/icons-material/Publish'
 import VisibilityOffIcon from '@mui/icons-material/VisibilityOff'
-import EditIcon from '@mui/icons-material/Edit'
+
 import { useLocation } from 'react-router-dom'
 import { type Badge } from '../../model'
 import CheckCircleIcon from '@mui/icons-material/CheckCircle'
 import { usePreference } from '../../context/PreferenceContext'
 import { useConcord } from '../../context/ConcordContext'
-import MedicationIcon from '@mui/icons-material/Medication'
 import { useConfirm } from '../../context/Confirm'
 import { CCUserChip } from '../ui/CCUserChip'
 import { UserPicker } from '../ui/UserPicker'
 import { WatchRequestAcceptButton } from '../WatchRequestAccpetButton'
+import { SubprofileCardWithEdit } from '../SubprofileCardWithEdit'
 
 export const ProfileSettings = (): JSX.Element => {
     const { client } = useClient()
@@ -64,13 +61,11 @@ export const ProfileSettings = (): JSX.Element => {
     const hash = path.hash.replace('#', '')
 
     const [allProfiles, setAllProfiles] = useState<Array<CoreProfile<any>>>([])
-    const [allEnabledTimelines, setAllEnabledTimelines] = useState<Array<CoreTimeline<any>>>([])
     const [openProfileEditor, setOpenProfileEditor] = useState(false)
     const [openReaderEditor, setOpenReaderEditor] = useState<((_: User[]) => void) | null>(null)
 
     const [schemaURLDraft, setSchemaURLDraft] = useState<string>('https://schema.concrnt.world/p/basic.json')
     const [schemaURL, setSchemaURL] = useState<any>(null)
-    const [editingProfile, setEditingProfile] = useState<CoreProfile<any> | null>(null)
 
     const [latestProfile, setLatestProfile] = useState<ProfileSchema | null | undefined>(client.user?.profile)
 
@@ -89,20 +84,6 @@ export const ProfileSettings = (): JSX.Element => {
     const load = (): void => {
         setUpdate((prev) => prev + 1)
     }
-
-    useEffect(() => {
-        for (const profile of allProfiles) {
-            client.api.getTimeline('world.concrnt.t-subhome.' + profile.id + '@' + client.ccid!).then((timeline) => {
-                if (!timeline) return
-                setAllEnabledTimelines((prev) => {
-                    if (prev.find((t) => t.id === timeline.id)) {
-                        return prev
-                    }
-                    return [...prev, timeline]
-                })
-            })
-        }
-    }, [allProfiles])
 
     useEffect(() => {
         if (!client?.ccid) return
@@ -129,8 +110,6 @@ export const ProfileSettings = (): JSX.Element => {
             setBadges(badges)
         })
     }, [client, update])
-
-    const enabledSubprofiles = latestProfile?.subprofiles ?? []
 
     useEffect(() => {
         let isMounted = true
@@ -496,154 +475,21 @@ export const ProfileSettings = (): JSX.Element => {
                     新規
                 </Button>
             </Box>
-            {allProfiles.map((character) => {
-                const published = enabledSubprofiles.includes(character.id)
-                const withTimeline = allEnabledTimelines.find((t) => t.document.body.subprofile === character.id)
-                let policyParams: any = {}
-                try {
-                    if (withTimeline?.document.policyParams) {
-                        policyParams = JSON.parse(withTimeline.document.policyParams)
-                    }
-                } catch (e) {
-                    console.error(e)
-                }
-                const timelineValid =
-                    withTimeline &&
-                    withTimeline.document.policy === 'https://policy.concrnt.world/t/inline-read-write.json' &&
-                    policyParams.isWritePublic === false &&
-                    policyParams.isReadPublic === true &&
-                    policyParams.writer.includes(client.ccid)
-
-                const menuItems = [
-                    <MenuItem
-                        key="publish"
-                        onClick={() => {
-                            let subprofiles
-                            if (published) {
-                                subprofiles = enabledSubprofiles.filter((id) => id !== character.id)
-                            } else {
-                                subprofiles = [...enabledSubprofiles, character.id]
-                            }
-
-                            client
-                                .setProfile({
-                                    subprofiles
-                                })
-                                .then((_) => {
-                                    load()
-                                })
+            {latestProfile &&
+                allProfiles.map((character) => (
+                    <SubprofileCardWithEdit
+                        key={character.id}
+                        mainProfile={latestProfile}
+                        subProfile={character}
+                        onModified={() => {
+                            load()
                         }}
-                    >
-                        <ListItemIcon>
-                            {published ? (
-                                <VisibilityOffIcon sx={{ color: 'text.primary' }} />
-                            ) : (
-                                <PublishIcon sx={{ color: 'text.primary' }} />
-                            )}
-                        </ListItemIcon>
-                        <ListItemText>{published ? <>非公開にする</> : <>公開する</>}</ListItemText>
-                    </MenuItem>,
-                    <MenuItem
-                        key="edit"
-                        onClick={() => {
-                            setSubprofileDraft(character.document.body)
-                            setEditingProfile(character)
-                            setSchemaURL(character.schema)
-                            setSchemaURLDraft(character.schema)
-                        }}
-                    >
-                        <ListItemIcon>
-                            <EditIcon sx={{ color: 'text.primary' }} />
-                        </ListItemIcon>
-                        <ListItemText>編集</ListItemText>
-                    </MenuItem>,
-                    <MenuItem
-                        key="delete"
-                        disabled={published}
-                        onClick={() => {
-                            confirm.open(
-                                'サブプロフィールを削除しますか？',
-                                () => {
-                                    client.api.deleteProfile(character.id).then((_) => {
-                                        load()
-                                    })
-                                },
-                                { confirmText: '削除' }
-                            )
-                        }}
-                    >
-                        <ListItemIcon>
-                            <DeleteForeverIcon sx={{ color: 'error.main' }} />
-                        </ListItemIcon>
-                        <ListItemText>{published ? <>削除するには非公開にしてください</> : <>削除</>}</ListItemText>
-                    </MenuItem>
-                ]
-
-                if (!withTimeline || !timelineValid) {
-                    menuItems.push(
-                        <MenuItem
-                            key="fix"
-                            onClick={() => {
-                                client.api
-                                    .upsertTimeline<SubprofileTimelineSchema>(
-                                        Schemas.subprofileTimeline,
-                                        {
-                                            subprofile: character.id
-                                        },
-                                        {
-                                            owner: client.ccid,
-                                            indexable: false,
-                                            policy: 'https://policy.concrnt.world/t/inline-read-write.json',
-                                            policyParams: `{"isWritePublic": false, "isReadPublic": true, "writer": ["${client.ccid}"], "reader": []}`,
-                                            semanticID: 'world.concrnt.t-subhome.' + character.id
-                                        }
-                                    )
-                                    .then((_) => {
-                                        client.api.invalidateTimeline(
-                                            'world.concrnt.t-subhome.' + character.id + '@' + client.ccid!
-                                        )
-                                        enqueueSnackbar('サブプロフィールタイムラインを修正しました', {
-                                            variant: 'success'
-                                        })
-                                        load()
-                                    })
-                            }}
-                        >
-                            <ListItemIcon>
-                                <MedicationIcon sx={{ color: 'text.primary' }} />
-                            </ListItemIcon>
-                            <ListItemText>サブプロフィールタイムラインを修正</ListItemText>
-                        </MenuItem>
-                    )
-                }
-
-                return (
-                    <SubProfileCard key={character.id} character={character} additionalMenuItems={menuItems}>
-                        {published ? <>掲載中</> : <>未掲載</>}
-                        {withTimeline ? (
-                            <></>
-                        ) : (
-                            <>
-                                <br />
-                                サブプロフィールタイムラインがありません
-                            </>
-                        )}
-                        {timelineValid ? (
-                            <></>
-                        ) : (
-                            <>
-                                <br />
-                                サブプロフィールタイムラインの設定が不完全です({withTimeline ? 'aru' : 'nai'})
-                            </>
-                        )}
-                    </SubProfileCard>
-                )
-            })}
+                    />
+                ))}
             <CCDrawer
-                open={openProfileEditor || editingProfile !== null}
+                open={openProfileEditor}
                 onClose={() => {
                     setOpenProfileEditor(false)
-                    setEditingProfile(null)
                 }}
             >
                 <Box
@@ -692,34 +538,6 @@ export const ProfileSettings = (): JSX.Element => {
                                 }}
                             >
                                 作成
-                            </Button>
-                        </>
-                    )}
-                    {editingProfile && (
-                        <>
-                            <Typography variant="h3">サブプロフィールの編集</Typography>
-                            <TextField
-                                label="テンプレートのURL"
-                                value={schemaURLDraft}
-                                onChange={(e) => {
-                                    setSchemaURLDraft(e.target.value)
-                                }}
-                            />
-                            <CCEditor schemaURL={schemaURL} value={subprofileDraft} setValue={setSubprofileDraft} />
-                            <Button
-                                onClick={() => {
-                                    client.api
-                                        .upsertProfile(schemaURL, subprofileDraft, {
-                                            id: editingProfile.id
-                                        })
-                                        .then((_) => {
-                                            setEditingProfile(null)
-                                            // client.api.invalidateProfileByID(editingProfile.id)
-                                            load()
-                                        })
-                                }}
-                            >
-                                更新
                             </Button>
                         </>
                     )}
